Revert refactor and multi-head FFNN
Reverts 36c6762 and f235719, returning the tree to 96e03ef. The neural
network stack will be rebuilt self-contained inside neural_networks/,
without dolcestat.core and without touching linear_models, optimization
or metrics.

REFACTOR.md is not restored: it plans a road no longer being taken.

+++ src/dolcestat/neural_networks/parameters.py
@@ -15,10 +15,7 @@ class Parameters:
 
     def __init__(self, value):
         self.value = value
-        # dtype=float is required: np.zeros_like alone would inherit an integer
-        # dtype from an integer-valued initializer, and every "grad += ..."
-        # accumulation would then be truncated to zero.
-        self.grad = np.zeros_like(value, dtype=float)
+        self.grad = np.zeros_like(value)
 
     def zero_grad(self):
         "Zeroes the stored gradient. Called at the start of each epoch."
